fix: split test_remote_backup_cli.py under 500 lines
Move S3-2 test classes (TestBackupDownloadFromServer,
TestBackupVerifyMetabase, TestBackupConfig) to new file
test_remote_backup_download.py. Both files now under 500 lines.
